refactor: combine MerchantButton and MerchantLink into MerchantAction
